requirements.txt
pyTelegramBotApi==4.29.1
python-dotenv==1.1.1